Use wkhtmltopdf directly (without extra node module)
When using the module sometimes the HTTP server would give 200 OK requests even when there was a failure.
This seemed to happen because ocassionaly the order of the events (write stream finished, child process exitted) would be flipped and the module would give an error when seeing this.
Also the order of sending the status code and sending output pdf needs to be kept.